🧹 [code health improvement] Refactor saveParser/index.ts into separate parser modules
Extracts the large saveParser/index.ts file into smaller, pure functions for each section of the save file.

Created src/engine/saveParser/parsers directory and extracted:
- common.ts: Types, interfaces, and shared utility functions.
- gen1.ts: Gen 1 parsing, structure checking, version detection, and mapping dictionaries.
- gen2.ts: Gen 2 parsing, structure checking, map dicts, and version detection.

Updated index.ts to import from these modular files, acting as a clean facade and maintaining exact original functionality without regressions.
